Standardize instrument construct
